Moved all View related commands from Lists to their own HelpCategory
