+++ Commands/Views/GetView.cs
@@ -6,11 +6,11 @@ using Microsoft.SharePoint.Client;
 using SharePointPnP.PowerShell.CmdletHelpAttributes;
 using SharePointPnP.PowerShell.Commands.Base.PipeBinds;
 
-namespace SharePointPnP.PowerShell.Commands.Lists
+namespace SharePointPnP.PowerShell.Commands.Views
 {
     [Cmdlet(VerbsCommon.Get, "PnPView")]
     [CmdletHelp("Returns one or all views from a list",
-        Category = CmdletHelpCategory.Lists,
+        Category = CmdletHelpCategory.Views,
         OutputType = typeof(View),
         OutputTypeLink = "https://msdn.microsoft.com/en-us/library/microsoft.sharepoint.client.view.aspx")]
     [CmdletExample(
