+++ Documentation/MSDN/Lists-category.md
@@ -16,6 +16,3 @@ Cmdlet|Description
 **[Set&#8209;PnPListPermission](SetPnPListPermission.md)** |Sets list permissions
 **[Get&#8209;PnPProvisioningTemplateFromGallery](GetPnPProvisioningTemplateFromGallery.md)** |Retrieves or searches provisioning templates from the PnP Template Gallery
 **[Request&#8209;PnPReIndexList](RequestPnPReIndexList.md)** |Marks the list for full indexing during the next incremental crawl
-**[Add&#8209;PnPView](AddPnPView.md)** |Adds a view to a list
-**[Get&#8209;PnPView](GetPnPView.md)** |Returns one or all views from a list
-**[Remove&#8209;PnPView](RemovePnPView.md)** |Deletes a view from a list
