+++ Documentation/MSDN/Listviews-category.md
@@ -1,0 +1,7 @@
+# List views 
+Cmdlet|Description
+:-----|:----------
+**[Add&#8209;PnPView](AddPnPView.md)** |Adds a view to a list
+**[Get&#8209;PnPView](GetPnPView.md)** |Returns one or all views from a list
+**[Remove&#8209;PnPView](RemovePnPView.md)** |Deletes a view from a list
+**[Set&#8209;PnPView](SetPnPView.md)** |Changes one or more properties of a specific view
