+++ Documentation/MSDN/TOC.md
@@ -66,7 +66,6 @@
 ### [Set-PnPField](SetPnPField.md)
 ### [Add-PnPFieldFromXml](AddPnPFieldFromXml.md)
 ### [Add-PnPTaxonomyField](AddPnPTaxonomyField.md)
-### [Set-PnPView](SetPnPView.md)
 ## [Files and Folders](FilesandFolders-category.md)
 ### [Add-PnPFile](AddPnPFile.md)
 ### [Copy-PnPFile](CopyPnPFile.md)
@@ -91,6 +90,11 @@
 ### [Set-PnPSiteClosure](SetPnPSiteClosure.md)
 ### [Set-PnPSitePolicy](SetPnPSitePolicy.md)
 ### [Get-PnPSitePolicy](GetPnPSitePolicy.md)
+## [List views](Listviews-category.md)
+### [Add-PnPView](AddPnPView.md)
+### [Get-PnPView](GetPnPView.md)
+### [Remove-PnPView](RemovePnPView.md)
+### [Set-PnPView](SetPnPView.md)
 ## [Lists](Lists-category.md)
 ### [Get-PnPDefaultColumnValues](GetPnPDefaultColumnValues.md)
 ### [Set-PnPDefaultColumnValues](SetPnPDefaultColumnValues.md)
@@ -107,9 +111,6 @@
 ### [Set-PnPListPermission](SetPnPListPermission.md)
 ### [Get-PnPProvisioningTemplateFromGallery](GetPnPProvisioningTemplateFromGallery.md)
 ### [Request-PnPReIndexList](RequestPnPReIndexList.md)
-### [Add-PnPView](AddPnPView.md)
-### [Get-PnPView](GetPnPView.md)
-### [Remove-PnPView](RemovePnPView.md)
 ## [Microsoft Graph](MicrosoftGraph-category.md)
 ### [Connect-PnPMicrosoftGraph](ConnectPnPMicrosoftGraph.md)
 ### [Get-PnPUnifiedGroup](GetPnPUnifiedGroup.md)
